# interface-specs/openapi/platform-api-v1.yaml
openapi: 3.0.3
info:
  title: 县域系统集成平台管理接口
  version: 0.1.0
  description: >-
    OpenAPI 是用机器可读文件说明 HTTP 接口的通用规范。本文件描述当前平台内部管理接口。
    业务接口只有在接口规范、已删除或遮盖敏感字段的样例、网络条件和负责人均确认后才能开发。
servers:
  - url: /api/v1
security:
  - hospitalIdentity: []
paths:
...
  /configuration/parameter-definitions:
    get:
      operationId: listParameterDefinitions
      summary: 查询代码中登记的平台参数定义
      responses:
        '200':
          description: 用于生成明确管理表单的已登记参数定义
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParameterDefinitionListResponse' }
        '403': { description: 'configuration:read permission is required' }
  /configuration/parameters:
    get:
      operationId: listParameterValues
      summary: 查询全局参数值和当前账号有权访问的机构参数值
      responses:
        '200':
          description: 当前管理员可查看的参数值
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParameterValueListResponse' }
        '403': { description: 'configuration:read permission is required' }
  /configuration/parameters/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 64, pattern: '^[a-z][a-z0-9_.-]*$' }
    put:
      operationId: upsertParameterValue
      summary: 新增或修改代码中已登记的参数值；如果期间已被他人修改则拒绝保存
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertParameterRequest' }
      responses:
        '200':
          description: 参数值已保存；敏感值仅返回固定遮盖符号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParameterValueResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission and organization scope are required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
    delete:
      operationId: deleteParameterValue
      summary: 删除当前环境和机构范围的参数值，使该范围恢复为尚未配置
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteParameterRequest' }
      responses:
        '200': { description: 当前范围的参数值已删除 }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission and organization scope are required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/dictionaries:
    get:
      operationId: listDictionaryTypes
      summary: 查询平台数据字典类型
      responses:
        '200':
          description: 平台数据字典类型
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryTypeListResponse' }
        '403': { description: 'configuration:read permission is required' }
    post:
      operationId: createDictionaryType
      summary: 创建平台数据字典类型
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDictionaryTypeRequest' }
      responses:
        '201':
          description: 已创建字典类型
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryTypeResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission is required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/dictionaries/{typeId}:
    parameters:
      - name: typeId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    put:
      operationId: updateDictionaryType
      summary: 修改或停用字典类型；如果期间已被他人修改则拒绝保存
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDictionaryTypeRequest' }
      responses:
        '200':
          description: 已修改字典类型
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryTypeResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
    delete:
      operationId: deleteDictionaryType
      summary: 删除不含任何字典项的字典类型
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteVersionRequest' }
      responses:
        '200': { description: 字典类型已删除 }
        '403': { description: 'configuration:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/dictionaries/{typeId}/items:
    parameters:
      - name: typeId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    get:
      operationId: listDictionaryItems
      summary: 查询指定字典类型中的字典项
      parameters:
        - name: includeDisabled
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: 按显示顺序和字典项代码排列的字典项
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryItemListResponse' }
        '403': { description: 'configuration:read permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
    post:
      operationId: createDictionaryItem
      summary: 向已启用的字典类型新增字典项
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDictionaryItemRequest' }
      responses:
        '201':
          description: 已创建字典项
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryItemResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/dictionary-items/{itemId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    put:
      operationId: updateDictionaryItem
      summary: 修改或停用字典项；如果期间已被他人修改则拒绝保存
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDictionaryItemRequest' }
      responses:
        '200':
          description: 已修改字典项
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryItemResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
    delete:
      operationId: deleteDictionaryItem
      summary: 删除未被业务数据引用的字典项
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteVersionRequest' }
      responses:
        '200': { description: 字典项已删除 }
        '403': { description: 'configuration:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/external-systems:
    get:
      operationId: listExternalSystems
      summary: 查询已确认的外部系统
      responses:
        '200': { description: 外部系统列表, content: { application/json: { schema: { $ref: '#/components/schemas/ExternalSystemListResponse' } } } }
        '403': { description: 'configuration:read permission is required' }
    post:
      operationId: createExternalSystem
      summary: 登记已确认的外部系统
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateExternalSystemRequest' } } } }
      responses:
        '201': { description: 已创建外部系统, content: { application/json: { schema: { $ref: '#/components/schemas/ExternalSystemResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission is required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/external-systems/{systemId}:
    parameters:
      - { name: systemId, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    put:
      operationId: updateExternalSystem
      summary: 修改或停用外部系统；如果期间已被他人修改则拒绝保存
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateExternalSystemRequest' } } } }
      responses:
        '200': { description: 已修改外部系统, content: { application/json: { schema: { $ref: '#/components/schemas/ExternalSystemResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/external-systems/{systemId}/endpoints:
    parameters:
      - { name: systemId, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    get:
      operationId: listExternalEndpoints
      summary: 查询全局服务地址和当前账号有权访问的机构服务地址
      responses:
        '200': { description: 不包含凭证引用的服务地址资料, content: { application/json: { schema: { $ref: '#/components/schemas/ExternalEndpointListResponse' } } } }
        '403': { description: 'configuration:read permission and organization scope are required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
    post:
      operationId: createExternalEndpoint
      summary: 创建默认停用的服务地址，待检查后再启用
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateExternalEndpointRequest' } } } }
      responses:
        '201': { description: 已创建默认停用的服务地址, content: { application/json: { schema: { $ref: '#/components/schemas/ExternalEndpointResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission and organization scope are required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /configuration/external-endpoints/{endpointId}:
    parameters:
      - { name: endpointId, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    put:
      operationId: updateExternalEndpoint
      summary: 修改或启用当前账号有权管理的服务地址；如果期间已被他人修改则拒绝保存
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateExternalEndpointRequest' } } } }
      responses:
        '200': { description: 不包含凭证引用的服务地址资料, content: { application/json: { schema: { $ref: '#/components/schemas/ExternalEndpointResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'configuration:write permission and organization scope are required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
...
components:
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: 由 GET /session/csrf 返回，并与 XSRF-TOKEN Cookie 配套使用的令牌
      schema: { type: string }
  securitySchemes:
    hospitalIdentity:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >-
        Cookie marker for an authenticated management session. The first development
        phase uses a server-side local-management session and secure HttpOnly cookie;
        production identity must be reviewed by the hospital before activation.
  responses:
    InvalidRequest:
      description: 请求参数校验未通过
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    ResourceNotFound:
      description: 请求的平台数据不存在，或当前账号无权查看
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ResourceConflict:
      description: 唯一值重复、当前状态不允许操作，或数据已被他人修改
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    DeleteVersionRequest:
      type: object
      required: [version]
      properties:
        version: { type: string, format: byte, description: 由 8 字节 SQL Server 行版本号编码得到的 Base64 文本 }
...
    ParameterDefinitionListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/ParameterDefinition' } }
...
    UpsertParameterRequest:
      type: object
      required: [environment, value, enabled]
      properties:
        environment: { type: string, enum: [DEVELOPMENT, TEST, PRODUCTION] }
        organizationId: { type: integer, format: int64, minimum: 1, nullable: true }
        value:
          type: string
          maxLength: 1000
          writeOnly: true
          description: 按代码中登记的参数定义校验；不得包含外部系统的明文凭证
        enabled: { type: boolean }
        version:
          type: string
          format: byte
          nullable: true
          description: 仅在首次创建该环境和机构对应的参数值时可以省略
    DeleteParameterRequest:
      type: object
      required: [environment, version]
      properties:
        environment: { type: string, enum: [DEVELOPMENT, TEST, PRODUCTION] }
        organizationId: { type: integer, format: int64, minimum: 1, nullable: true }
        version: { type: string, format: byte, description: 由 8 字节 SQL Server 行版本号编码得到的 Base64 文本 }
    ParameterValueListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/ParameterValue' } }
    ParameterValueResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { $ref: '#/components/schemas/ParameterValue' }
...
    DictionaryTypeListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/DictionaryType' } }
    DictionaryTypeResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { $ref: '#/components/schemas/DictionaryType' }
...
    CreateDictionaryTypeRequest:
      type: object
      required: [typeCode, typeName, description]
      properties:
        typeCode: { type: string, minLength: 1, maxLength: 64, pattern: '^[A-Za-z][A-Za-z0-9_.-]*$' }
        typeName: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, minLength: 1, maxLength: 500 }
    UpdateDictionaryTypeRequest:
      type: object
      required: [typeName, description, enabled, version]
      properties:
        typeName: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, minLength: 1, maxLength: 500 }
        enabled: { type: boolean }
        version: { type: string, format: byte }
    DictionaryItemListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/DictionaryItem' } }
    DictionaryItemResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { $ref: '#/components/schemas/DictionaryItem' }
...
    CreateDictionaryItemRequest:
      type: object
      required: [itemCode, itemLabel, sortOrder]
      properties:
        itemCode: { type: string, minLength: 1, maxLength: 64, pattern: '^[A-Za-z][A-Za-z0-9_.-]*$' }
        itemLabel: { type: string, minLength: 1, maxLength: 200 }
        sortOrder: { type: integer, minimum: 0, maximum: 999999 }
    UpdateDictionaryItemRequest:
      type: object
      required: [itemLabel, sortOrder, enabled, version]
      properties:
        itemLabel: { type: string, minLength: 1, maxLength: 200 }
        sortOrder: { type: integer, minimum: 0, maximum: 999999 }
        enabled: { type: boolean }
        version: { type: string, format: byte }
    ExternalSystemListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/ExternalSystem' } }
    ExternalSystemResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { $ref: '#/components/schemas/ExternalSystem' }
...
    CreateExternalSystemRequest:
      type: object
      required: [systemCode, systemName, description]
      properties:
        systemCode: { type: string, minLength: 1, maxLength: 64, pattern: '^[A-Za-z][A-Za-z0-9_.-]*$' }
        systemName: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, minLength: 1, maxLength: 500 }
    UpdateExternalSystemRequest:
      type: object
      required: [systemName, description, enabled, version]
      properties:
        systemName: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, minLength: 1, maxLength: 500 }
        enabled: { type: boolean }
        version: { type: string, format: byte }
    ExternalEndpointListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/ExternalEndpoint' } }
    ExternalEndpointResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { $ref: '#/components/schemas/ExternalEndpoint' }
...
    ApiError:
      type: object
      required: [code, message, timestamp]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }